requirements.txt
numpy~=1.18.5
scipy==1.4.1
tensorflow~=2.3.0
tensorflow_text>=2.0.0rc0
tensorflow_hub
tensorflow-datasets
pandas~=1.0.5
elasticsearch~=7.8.0
pyfunctional~=1.4.3
